+++ lib/cleanup.php
@@ -191,7 +191,6 @@ function roots_embed_wrap($cache, $url, $attr = '', $post_ID = '') {
   return '<div class="entry-content-asset">' . $cache . '</div>';
 }
 add_filter('embed_oembed_html', 'roots_embed_wrap', 10, 4);
-add_filter('embed_googlevideo', 'roots_embed_wrap', 10, 2);
 
 /**
  * Add class="thumbnail" to attachment items
@@ -385,61 +384,6 @@ function roots_remove_default_description($bloginfo) {
   return ($bloginfo === $default_tagline) ? '' : $bloginfo;
 }
 add_filter('get_bloginfo_rss', 'roots_remove_default_description');
-
-/**
- * Allow more tags in TinyMCE including <iframe> and <script>
- */
-function roots_change_mce_options($options) {
-  $ext = 'pre[id|name|class|style],iframe[align|longdesc|name|width|height|frameborder|scrolling|marginheight|marginwidth|src],script[charset|defer|language|src|type]';
-
-  if (isset($initArray['extended_valid_elements'])) {
-    $options['extended_valid_elements'] .= ',' . $ext;
-  } else {
-    $options['extended_valid_elements'] = $ext;
-  }
-
-  return $options;
-}
-add_filter('tiny_mce_before_init', 'roots_change_mce_options');
-
-/**
- * Add additional classes onto widgets
- *
- * @link http://wordpress.org/support/topic/how-to-first-and-last-css-classes-for-sidebar-widgets
- */
-function roots_widget_first_last_classes($params) {
-  global $my_widget_num;
-
-  $this_id = $params[0]['id'];
-  $arr_registered_widgets = wp_get_sidebars_widgets();
-
-  if (!$my_widget_num) {
-    $my_widget_num = array();
-  }
-
-  if (!isset($arr_registered_widgets[$this_id]) || !is_array($arr_registered_widgets[$this_id])) {
-    return $params;
-  }
-
-  if (isset($my_widget_num[$this_id])) {
-    $my_widget_num[$this_id] ++;
-  } else {
-    $my_widget_num[$this_id] = 1;
-  }
-
-  $class = 'class="widget-' . $my_widget_num[$this_id] . ' ';
-
-  if ($my_widget_num[$this_id] == 1) {
-    $class .= 'widget-first ';
-  } elseif ($my_widget_num[$this_id] == count($arr_registered_widgets[$this_id])) {
-    $class .= 'widget-last ';
-  }
-
-  $params[0]['before_widget'] = preg_replace('/class=\"/', "$class", $params[0]['before_widget'], 1);
-
-  return $params;
-}
-add_filter('dynamic_sidebar_params', 'roots_widget_first_last_classes');
 
 /**
  * Redirects search results from /?s=query to /search/query/, converts %20 to +